requirements.txt
django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
qrcode==7.4.2
Pillow==10.4.0

django-filter==24.3
drf-spectacular==0.27.2
firebase-admin==6.5.0
cryptography==43.0.0
stripe==8.9.0

